buttons moved to right of canvas. Text display still being figured out.
new error in lib code -press stop, continue to get around when starting 
working on fix currently but not sure how to approach code in the libraries. 
I think it might be based on an overlap of variable names between existing code and the library but still working on solutions.